fix: BrowserChrome service location, Cookie catching
* update BrowserChrome selenium call to include local chromedriver.
We are already packaging chromedriver.exe in with ghosts, so this
just references an executable we already have. This prevents
selenium from trying to reach out to the web and check for a newer
chromedriver version, which prevented selenium from working in air-
gapped (cyber range) environments.
* update BaseBrowser so POST/PUT/DELETE actions browse to the site
in question before executing thier action. This allows for combined
browsing and HTTP actions, so that websites can be logged into and
then browsed (ref https://github.com/cmu-sei/GHOSTS/discussions/517)

Other:
  * update NLog version in Ghosts.Client.Windows to match Domain
  * update build_windows.ps1 to allow multiple msbuild.exe locations